format: Auto-format code by ruff [skip actions].

+++ bot/modules/services.py
@@ -72,9 +72,9 @@ async def start(client, message):
 
 @new_task
 async def ping(_, message):
-    start_time = int(round(time() * 1000))
+    start_time = round(time() * 1000)
     reply = await send_message(message, "Starting Ping")
-    end_time = int(round(time() * 1000))
+    end_time = round(time() * 1000)
     await edit_message(reply, f"{end_time - start_time} ms")
